// frontend/src/app/routes/organigrama/organigramaRamas_Subramas/types/rama.type.ts
export interface Rama {
  section_id: string;
  sectionName: string;
  sectionDescription?: string;
  sectionGalleryObjectIds: string[];
  id: string;
  nombre: string;
  descripcion?: string;
  icono?: string;
  iconoObjectId?: string; // ID para la imagen en localStorage
  imagenPrincipal?: string; // URL de la imagen principal
  imagenPrincipalObjectId?: string; // ID para la imagen principal en localStorage
  edadMinima: number;
  edadMaxima: number;
  año: number;
  estado: 'activa' | 'inactiva';
  fechaCreacion: string;
  subramas: Subrama[];
}

export interface Subrama {
  subgroup_id: string; 
  subgroupName: string;
  subgroupDescription?: string;
  subgroupGalleryObjectIds?: string[]; // IDs de imágenes de galería en localStorage
  section_id: string; 
  id: string; 
  nombre: string;
  descripcion?: string;
  icono?: string;
  iconoObjectId?: string; // ID para la imagen en localStorage
  imagenPrincipal?: string; // URL de la imagen principal
  imagenPrincipalObjectId?: string; // ID para la imagen principal en localStorage
  ramaId: string;
  lider?: string;
  estado: 'activa' | 'inactiva';
  fechaCreacion: string;
  numeroMiembros: number;
}